Fixes the bug with upgrading engine version
Recently we were encountering the described error here:
https://github.com/terraform-aws-modules/terraform-aws-rds-aurora/issues/118
when we were upgrading from 10.11 to 10.14. It dropped the instances and then stopped right in the middle with a 400 - leaving everything in pieces behind :(.
I tried the fix locally by modifying the `terraform/apps/oma/.terraform/eu01/stg02-de/modules/oma.db/main.tf` adding the changes and it seems to work as expected.